+++ fig.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-cp files/auth.json dockerfiles/ezpublish/prepare
+if [ -f files/auth.json ]; then
+    echo bla1
+    cp files/auth.json dockerfiles/ezpublish/prepare
+    echo bla2
+else
+    touch dockerfiles/ezpublish/prepare/auth.json
+fi
 cp files/etcd_functions dockerfiles/etcd
 cp files/etcd_functions dockerfiles/mysql
 cp files/etcd_functions dockerfiles/php-fpm
